backend/file-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pymongo==4.3.3
python-dotenv==1.0.0
